ci_template: avoid building libnice tests and examples in the subproject
We don't run the libnice testsuite, and when binaries are built
they consume ~45mb of space. This increases the size of the artifacts
we export from the gst-build job for the testsuite and drives up
the storage and bandwith costs when re-downloading the artifacts.

Similary disable the test targets of couple other subprojects as well

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-ci/-/merge_requests/283>